Overhauling how XNA/Monogame Content is being handled in the texture loader, namely as backup.
