+++ SKSSL/SKSSL.Engine.csproj
@@ -114,6 +114,9 @@
         <Compile Update="Textures\TextureLoader.IconRegistry.cs">
           <DependentUpon>TextureLoader.cs</DependentUpon>
         </Compile>
+        <Compile Update="Textures\TextureLoader.XNAContent.cs">
+          <DependentUpon>TextureLoader.cs</DependentUpon>
+        </Compile>
     </ItemGroup>
 
     <!-- !== ADD THE FOLLOWING TO GAME PROJECTS INHERITING / USING SKSSL AS AN ENGINE !== -->
